scanner: Refactor scanners for more modularity
Move scanDependenciesFile() to the Scanner class and scanInputPath() to
the LocalScanner class, respectively. To avoid passing the
ScannerConfiguration to both functions, make it an argument of the
Scanner constructor. As a result, Scanners cannot directly be
instantiated via ServiceLoader anymore which requires a public no-argument
constructor. Thus wrapping Scanners in new AbstractScannerFactory classes
that implement a common ScannerFactory interface becomes necessary.
