+++ scanner/src/main/kotlin/Scanner.kt
@@ -19,15 +19,24 @@
 
 package com.here.ort.scanner
 
+import com.here.ort.downloader.consolidateProjectPackagesByVcs
+import com.here.ort.model.Environment
+import com.here.ort.model.OrtResult
 import com.here.ort.model.Package
+import com.here.ort.model.ProjectScanScopes
+import com.here.ort.model.ScanRecord
 import com.here.ort.model.ScanResult
+import com.here.ort.model.ScanResultContainer
+import com.here.ort.model.ScannerRun
+import com.here.ort.model.config.ScannerConfiguration
+import com.here.ort.model.readValue
 
 import java.io.File
 import java.util.ServiceLoader
 
-abstract class Scanner {
+abstract class Scanner(protected val config: ScannerConfiguration) {
     companion object {
-        private val LOADER by lazy { ServiceLoader.load(Scanner::class.java)!! }
+        private val LOADER by lazy { ServiceLoader.load(ScannerFactory::class.java)!! }
 
         /**
          * The list of all available scanners in the classpath.
@@ -36,7 +45,7 @@ abstract class Scanner {
     }
 
     /**
-     * Return the Java class name as a simple way to refer to the scanner.
+     * Return the Java class name as a simple way to refer to the [Scanner].
      */
     override fun toString(): String = javaClass.simpleName
 
@@ -48,4 +57,74 @@ abstract class Scanner {
      */
     abstract fun scan(packages: List<Package>, outputDirectory: File, downloadDirectory: File? = null)
             : Map<Package, List<ScanResult>>
+
+    /**
+     * Scan the [Project]s and [Package]s specified in [dependenciesFile] using this [Scanner] and store the scan
+     * results in [outputDirectory]. If [downloadDirectory] is specified, it is used instead of [outputDirectory] to
+     * download the source code to. Return scan results as an [OrtResult].
+     */
+    fun scanDependenciesFile(dependenciesFile: File, outputDirectory: File, downloadDirectory: File? = null,
+                             scopesToScan: List<String> = emptyList()): OrtResult {
+        require(dependenciesFile.isFile) {
+            "Provided path for the configuration does not refer to a file: ${dependenciesFile.absolutePath}"
+        }
+
+        val ortResult = dependenciesFile.readValue(OrtResult::class.java)
+
+        require(ortResult.analyzer != null) {
+            "The provided dependencies file '${dependenciesFile.invariantSeparatorsPath}' does not contain an " +
+                    "analyzer result."
+        }
+
+        val analyzerResult = ortResult.analyzer!!.result
+
+        // Add the projects as packages to scan.
+        val consolidatedProjectPackageMap = consolidateProjectPackagesByVcs(analyzerResult.projects)
+        val consolidatedReferencePackages = consolidatedProjectPackageMap.keys.map { it.toCuratedPackage() }
+
+        val projectScanScopes = if (scopesToScan.isNotEmpty()) {
+            println("Limiting scan to scopes: $scopesToScan")
+
+            analyzerResult.projects.map { project ->
+                project.scopes.map { it.name }.partition { it in scopesToScan }.let {
+                    ProjectScanScopes(project.id, it.first.toSortedSet(), it.second.toSortedSet())
+                }
+            }
+        } else {
+            analyzerResult.projects.map {
+                val scopes = it.scopes.map { it.name }
+                ProjectScanScopes(it.id, scopes.toSortedSet(), sortedSetOf())
+            }
+        }.toSortedSet()
+
+        val packagesToScan = if (scopesToScan.isNotEmpty()) {
+            consolidatedReferencePackages + analyzerResult.packages.filter { pkg ->
+                analyzerResult.projects.any { it.scopes.any { it.name in scopesToScan && pkg.pkg in it } }
+            }
+        } else {
+            consolidatedReferencePackages + analyzerResult.packages
+        }.toSortedSet()
+
+        val results = scan(packagesToScan.map { it.pkg }, outputDirectory, downloadDirectory)
+        val resultContainers = results.map { (pkg, results) ->
+            // Remove the raw results from the scan results to reduce the size of the scan result.
+            // TODO: Consider adding an option to keep the raw results.
+            ScanResultContainer(pkg.id, results.map { it.copy(rawResult = null) })
+        }.toSortedSet()
+
+        // Add scan results from de-duplicated project packages to result.
+        consolidatedProjectPackageMap.forEach { referencePackage, deduplicatedPackages ->
+            resultContainers.find { it.id == referencePackage.id }?.let { resultContainer ->
+                deduplicatedPackages.forEach {
+                    resultContainers += resultContainer.copy(id = it.id)
+                }
+            }
+        }
+
+        val scanRecord = ScanRecord(projectScanScopes, resultContainers, ScanResultsCache.stats)
+
+        val scannerRun = ScannerRun(Environment(), config, scanRecord)
+
+        return OrtResult(ortResult.repository, ortResult.analyzer, scannerRun)
+    }
 }
