+++ scanner/src/main/kotlin/scanners/BoyterLc.kt
@@ -29,10 +29,12 @@ import com.here.ort.model.Provenance
 import com.here.ort.model.ScanResult
 import com.here.ort.model.ScanSummary
 import com.here.ort.model.ScannerDetails
+import com.here.ort.model.config.ScannerConfiguration
 import com.here.ort.model.jsonMapper
 import com.here.ort.scanner.LocalScanner
 import com.here.ort.scanner.Main
 import com.here.ort.scanner.ScanException
+import com.here.ort.scanner.AbstractScannerFactory
 import com.here.ort.utils.OkHttpClientHelper
 import com.here.ort.utils.OS
 import com.here.ort.utils.ProcessCapture
@@ -49,7 +51,11 @@ import okhttp3.Request
 
 import okio.Okio
 
-class BoyterLc : LocalScanner() {
+class BoyterLc(config: ScannerConfiguration) : LocalScanner(config) {
+    class Factory : AbstractScannerFactory<BoyterLc>() {
+        override fun create(config: ScannerConfiguration) = BoyterLc(config)
+    }
+
     override val scannerExe = if (OS.isWindows) "lc.exe" else "lc"
     override val scannerVersion = "1.3.1"
     override val resultFileExt = "json"
