+++ scanner/src/main/kotlin/scanners/FileCounter.kt
@@ -26,8 +26,10 @@ import com.here.ort.model.Provenance
 import com.here.ort.model.ScanResult
 import com.here.ort.model.ScanSummary
 import com.here.ort.model.ScannerDetails
+import com.here.ort.model.config.ScannerConfiguration
 import com.here.ort.model.jsonMapper
 import com.here.ort.scanner.LocalScanner
+import com.here.ort.scanner.AbstractScannerFactory
 
 import java.io.File
 import java.time.Instant
@@ -37,7 +39,11 @@ import java.time.Instant
  * scanners it is useful for testing the scanner tool, for example during development or when integrating it with other
  * tools.
  */
-class FileCounter : LocalScanner() {
+class FileCounter(config: ScannerConfiguration) : LocalScanner(config) {
+    class Factory : AbstractScannerFactory<FileCounter>() {
+        override fun create(config: ScannerConfiguration) = FileCounter(config)
+    }
+
     data class FileCountResult(val fileCount: Int)
 
     override val resultFileExt = "json"
