+++ scanner/src/main/kotlin/scanners/ScanCode.kt
@@ -31,9 +31,11 @@ import com.here.ort.model.Provenance
 import com.here.ort.model.ScanResult
 import com.here.ort.model.ScanSummary
 import com.here.ort.model.ScannerDetails
+import com.here.ort.model.config.ScannerConfiguration
 import com.here.ort.model.jsonMapper
 import com.here.ort.scanner.LocalScanner
 import com.here.ort.scanner.ScanException
+import com.here.ort.scanner.AbstractScannerFactory
 import com.here.ort.utils.OS
 import com.here.ort.utils.ProcessCapture
 import com.here.ort.utils.getCommandVersion
@@ -52,7 +54,11 @@ import java.util.SortedSet
 
 import kotlin.math.absoluteValue
 
-class ScanCode : LocalScanner() {
+class ScanCode(config: ScannerConfiguration) : LocalScanner(config) {
+    class Factory : AbstractScannerFactory<ScanCode>() {
+        override fun create(config: ScannerConfiguration) = ScanCode(config)
+    }
+
     private val OUTPUT_FORMAT = "json-pp"
     private val TIMEOUT = 300
