--- scanner/src/main/resources/META-INF/services/com.here.ort.scanner.Scanner
@@ -1,5 +1,0 @@
-com.here.ort.scanner.scanners.Askalono
-com.here.ort.scanner.scanners.BoyterLc
-com.here.ort.scanner.scanners.FileCounter
-com.here.ort.scanner.scanners.Licensee
-com.here.ort.scanner.scanners.ScanCode
